Update run-rt-forecast.R
EpiNow2 needs some level of non-zero data in the time-series in order to produce sane results. This check will cause estimates to not be run in places without at least 14 days of non-zero in the last 12 weeks (or fewer if using a different rolling window to the global estimates). Not sure but this may require some additional downstream checks as I expect the formatting/ensembling code wants there to always be forecasts.